% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Plot Population Growth for a Given Country}
\usage{
CountryPopulation(country_name)
}
\arguments{
\item{country_name}{The name of the country for which the population growth plot is desired.}
}
\value{
A ggplot object showing the population trend.
}
\description{
This function takes a country name as input and returns a ggplot object displaying the population growth from 1950 to 2020.
The plot is in black and white, with the country's name at the top as the title.
}
\examples{
CountryPopulation("Brazil")
}
