% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population Data}
\format{
A data frame with 271 rows and 72 columns.
}
\source{
United Nations Population Estimates.
}
\usage{
WorldPopulation
}
\description{
This dataset contains cleaned population data from 1950 to 2020 for UN countries,
filtered to exclude regional and income group aggregates. The data was sourced from
the United Nations database and processed to include only the total population estimates
for each year.
}
\keyword{datasets}
